requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
fake-headers==1.0.2
html5lib==1.1
idna==2.10
multiprocessing-logging==0.3.1
psycopg2-binary==2.8.6
PySocks==1.7.1
redis==3.5.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
webencodings==0.5.1
